+++ app/mailers/application_mailer.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 class ApplicationMailer < Mail::Notify::Mailer
+  include ApplicationHelper
 end