+++ CHANGELOG.md
@@ -1061,6 +1061,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 ## [unreleased]
 
 - Display a banner on non-production environments to make it clearer to users which site they're using
+- Show the environment name (e.g. "training") in the subject of emails sent by the application
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-112...HEAD
 [release-112]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-111...release-112